# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy Management System API",
        "description": "Complete API documentation for Pharmacy Management System including admin, student, teacher, and payment modules",
        "contact": {
            "email": "support@pharmacy.edu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.pharmacy.edu",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/admin/subjects": {
            "post": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Get all subjects for a department",
                "description": "Retrieves all subjects for a specific department, semester, and subject category. Note: semester_id should be numeric string (1, 2, 3) from semester dropdown's semesterId field. Calls: fn_admin_getdeptallsubjects_v1",
                "operationId": "0f3fd6e83a2378a7d1827cdc9b933744",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "admin_user_id",
                                    "dept_code",
                                    "semester_id",
                                    "subject_category_id"
                                ],
                                "properties": {
                                    "admin_user_id": {
                                        "description": "Admin user ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dept_code": {
                                        "description": "Department code (e.g., PHARM, BIO)",
                                        "type": "string",
                                        "example": "PHARM"
                                    },
                                    "semester_id": {
                                        "description": "Semester ID as numeric string (1-6)",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "subject_category_id": {
                                        "description": "Subject category ID (1=Theory, 2=Practical, etc.)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subjects retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully retrieved subjects."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "subjectId": {
                                                        "type": "integer",
                                                        "example": 101
                                                    },
                                                    "subjectCode": {
                                                        "type": "string",
                                                        "example": "PHARM101"
                                                    },
                                                    "subjectName": {
                                                        "type": "string",
                                                        "example": "Pharmaceutical Chemistry I"
                                                    },
                                                    "deptCode": {
                                                        "type": "string",
                                                        "example": "PHARM"
                                                    },
                                                    "semesterId": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "subjectCategoryId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Enter your authentication token",
                "name": "token",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}